+++ src/progress.tsx
@@ -1,10 +1,25 @@
-import { createContext, useCallback, useMemo, useState } from 'react';
+import { createContext, useCallback, useMemo, useRef, useState } from 'react';
 
 type Progress = {
   task: string;
   screen: string;
+
+  /**
+   * Like `updateTask`, but does nothing if progress has been restored previously.
+   * If `tryRestoringProgress` hasn't been called, calling `initTask` will only update
+   * local progress if `tryRestoringProgress` fails.
+   */
+  initTask: (newTask: string) => void;
+  /**
+   * Similar to `initScreen`.
+   */
+  initScreen: (newScreen: string) => void;
+  /**
+   * Calling updateTask with '' will clear local progress.
+   */
   updateTask: (newTask: string) => void;
   updateScreen: (newScreen: string) => void;
+
   /**
    * Attempts to restore local progress.
    *
@@ -12,11 +27,15 @@ type Progress = {
    * seconds. Otherwise, any existing progress will be cleared and the provided
    * `userId` will be associated with new progress saved.
    *
+   * Must be called before `updateTask` or `updateScreen` to set the user ID.
+   * Otherwise, a later call to `tryRestoringProgress` will not be able to restore
+   * progress, since the user ID will not match.
+   *
    * @param userId user ID to restore local progress for
    * @param maxAge max age of local progress (in seconds) to restore
-   * @returns the ID of the restored task, or '' if no progress was restored
+   * @returns the IDs of the restored task and screen, or ['', ''] if no progress was restored
    */
-  tryRestoringProgress: (userId: string, maxAge: number) => string;
+  tryRestoringProgress: (userId: string, maxAge: number) => [string, string];
 };
 
 const PROGRESS_ITEM_USERID = 'prog_userId';
@@ -27,6 +46,8 @@ const PROGRESS_ITEM_SCREEN = 'prog_screen';
 const ProgressDefault: Progress = {
   task: '',
   screen: '',
+  initTask: () => { throw new Error('Progress context not found.'); },
+  initScreen: () => { throw new Error('Progress context not found.'); },
   updateTask: () => { throw new Error('Progress context not found.'); },
   updateScreen: () => { throw new Error('Progress context not found.'); },
   tryRestoringProgress: () => { throw new Error('Progress context not found.'); },
@@ -40,10 +61,25 @@ const ProgressContext = createContext<Progress>(ProgressDefault);
  * Provides a ProgressContext.
  */
 function ProgressContextProvider({ children }: { children: React.ReactNode }) {
+  // Invariants:
+  // - PROGRESS_ITEM_TIMESTAMP will be set iff PROGRESS_ITEM_TASK is.
+  // - PROGRESS_ITEM_TASK cannot be ''.
   const [task, setTask] = useState('');
   const [screen, setScreen] = useState('');
+  /**
+   * Ref for communicating between `tryRestoringProgress` and `initTask`/`initScreen`.
+   * - If there is pending progress to save from `initTask`/`initScreen`, this value
+   * will be a tuple of the task and screen IDs.
+   * - If `tryRestoringProgress` has already been called and failed, this value will
+   * be `true`.
+   * - If progress has been restored, this value will be `false`.
+   */
+  const deferredProgressRef = useRef<boolean | [string, string]>(['', '']);
 
-  const tryRestoringProgress = useCallback((userId: string, maxAge: number) => {
+  const tryRestoringProgress = useCallback((userId: string, maxAge: number): [string, string] => {
+    if (typeof deferredProgressRef.current === 'boolean') {
+      throw new Error('tryRestoringProgress should not be called twice.');
+    }
     if (maxAge <= 0) {
       throw new Error('maxAge must be positive.');
     }
@@ -54,45 +90,89 @@ function ProgressContextProvider({ children }: { children: React.ReactNode }) {
       const storedScreen = localStorage.getItem(PROGRESS_ITEM_SCREEN);
       setTask(storedTask!);
       setScreen(storedScreen!);
+      deferredProgressRef.current = false;
       // careful, we're ignoring the possibility of null here
-      return (storedTask as string);
+      return [storedTask as string, storedScreen as string];
     }
     else {
-      // remove stale local progress
       localStorage.setItem(PROGRESS_ITEM_USERID, userId);
-      localStorage.removeItem(PROGRESS_ITEM_TIMESTAMP);
+      const [deferredTask, deferredScreen] = deferredProgressRef.current as [string, string];
+      // clear stale local progress
       localStorage.removeItem(PROGRESS_ITEM_TASK);
       localStorage.removeItem(PROGRESS_ITEM_SCREEN);
-      return '';
+      localStorage.removeItem(PROGRESS_ITEM_TIMESTAMP);
+      if (deferredTask !== '') {
+        updateTask(deferredTask);
+      }
+      if (deferredScreen !== '') {
+        updateScreen(deferredScreen);
+      }
+      deferredProgressRef.current = true;
+      return ['', ''];
     }
-  }, []);
+  }, [deferredProgressRef]);
+
+  const initTask = useCallback((newTask: string) => {
+    console.log(`Init task called, deferred task is ${deferredProgressRef.current}.`);
+    if (deferredProgressRef.current === false) {
+      // tryRestoringProgress succeeded
+      return;
+    }
+    if (deferredProgressRef.current === true) {
+      updateTask(newTask);
+    }
+    else {
+      console.log('Here, new task is ' + newTask);
+      setTask(newTask);
+      deferredProgressRef.current[0] = newTask;
+    }
+  }, [deferredProgressRef]);
+
+  const initScreen = useCallback((newScreen: string) => {
+    if (deferredProgressRef.current === false) {
+      // tryRestoringProgress succeeded
+      return;
+    }
+    if (deferredProgressRef.current === true) {
+      updateScreen(newScreen);
+    }
+    else {
+      setScreen(newScreen);
+      deferredProgressRef.current[1] = newScreen;
+    }
+  }, [deferredProgressRef]);
 
   const updateTask = useCallback((newTask: string) => {
-    setTask(newTask);
     if (newTask === '') {
-      localStorage.removeItem(PROGRESS_ITEM_TIMESTAMP);
-      localStorage.removeItem(PROGRESS_ITEM_TASK);
-      localStorage.removeItem(PROGRESS_ITEM_SCREEN);
+      throw new Error('Cannot update task to empty.');
     }
-    else {
-      localStorage.setItem(PROGRESS_ITEM_TASK, newTask);
-      localStorage.setItem(PROGRESS_ITEM_TIMESTAMP, Date.now().toString());
-    }
+    setTask(newTask);
+    localStorage.setItem(PROGRESS_ITEM_TASK, newTask);
+    localStorage.setItem(PROGRESS_ITEM_TIMESTAMP, Date.now().toString());
   }, []);
 
   const updateScreen = useCallback((newScreen: string) => {
+    if (newScreen === '') {
+      throw new Error('Cannot update screen to empty.');
+    }
     setScreen(newScreen);
     localStorage.setItem(PROGRESS_ITEM_SCREEN, newScreen);
-    localStorage.setItem(PROGRESS_ITEM_TIMESTAMP, Date.now().toString());
+    // only update timestamp if a task has been selected
+    const storedTask = localStorage.getItem(PROGRESS_ITEM_TASK);
+    if (storedTask) {
+      localStorage.setItem(PROGRESS_ITEM_TIMESTAMP, Date.now().toString());
+    }
   }, []);
 
   const progress = useMemo(() => ({
     task,
     screen,
+    initTask,
+    initScreen,
     updateTask,
     updateScreen,
     tryRestoringProgress,
-  }), [task, screen, updateTask, updateScreen, tryRestoringProgress]);
+  }), [task, screen, initTask, initScreen, updateTask, updateScreen, tryRestoringProgress]);
 
   return (
     <ProgressContext.Provider value={progress}>